requirements.txt
streamlit>=1.33.0
opencv-python>=4.9.0.80
Pillow>=10.3.0
numpy>=1.26.4
chromadb>=0.5.0
sentence-transformers>=2.7.0
transformers>=4.41.1
torch>=2.3.0
torchvision>=0.18.0
faster-whisper>=1.0.0
moviepy>=1.0.3
tqdm>=4.66.4
pandas>=2.2.2
soundfile>=0.12.1
ffmpeg-python>=0.2.0
accelerate>=0.31.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
